move db attribute normalization and unit tests into the eap normalization crate and add function documentation
