+++ relay-server/src/processing/spans/process.rs
@@ -5,19 +5,12 @@ use crate::managed::Managed;
 use crate::processing::Context;
 use crate::processing::spans::{self, Error, ExpandedSpans, Result, SampledSpans};
 use crate::services::outcome::DiscardReason;
-use relay_conventions::consts;
-use relay_event_normalization::span::TABLE_NAME_REGEX;
-use relay_event_normalization::span::description::scrub_db_query;
-use relay_event_normalization::span::tag_extraction::{
-    sql_action_from_query, sql_tables_from_query,
-};
 use relay_event_normalization::{
     GeoIpLookup, RequiredMode, SchemaProcessor, TimestampProcessor, TrimmingProcessor, eap,
 };
 use relay_event_schema::processor::{ProcessingState, ValueType, process_value};
-use relay_event_schema::protocol::{Attributes, Span, SpanV2};
+use relay_event_schema::protocol::{Span, SpanV2};
 use relay_protocol::Annotated;
-use std::borrow::Cow;
 
 /// Parses all serialized spans.
 ///
@@ -116,7 +109,7 @@ fn normalize_span(
         }
         eap::normalize_ai(&mut span.attributes, duration, model_costs);
 
-        normalize_attribute_values(&mut span.attributes);
+        eap::normalize_attribute_values(&mut span.attributes);
     };
 
     process_value(span, &mut TrimmingProcessor::new(), ProcessingState::root())?;
@@ -132,131 +125,6 @@ fn normalize_span(
     }
 
     Ok(())
-}
-
-fn normalize_attribute_values(attributes: &mut Annotated<Attributes>) {
-    normalize_db_attributes(attributes);
-}
-
-fn normalize_db_attributes(attributes: &mut Annotated<Attributes>) {
-    if attributes
-        .value()
-        .is_some_and(|attributes| attributes.get_value(consts::NORMALIZED_DB_QUERY).is_some())
-    {
-        return;
-    }
-
-    let (normalized_db_query, db_operation, db_collection_name) = attributes
-        .value()
-        .map(|attributes| {
-            let (op, sub_op) = attributes
-                .get_value(consts::OP)
-                .and_then(|v| v.as_str())
-                .and_then(|op| op.split_once('.'))
-                .unwrap_or_default();
-
-            let raw_query = attributes
-                .get_value(consts::DB_QUERY_TEXT)
-                .or_else(|| {
-                    if op == "db" {
-                        attributes.get_value(consts::DESCRIPTION)
-                    } else {
-                        None
-                    }
-                })
-                .and_then(|v| v.as_str());
-
-            let db_system = attributes
-                .get_value(consts::DB_SYSTEM_NAME)
-                .and_then(|v| v.as_str());
-
-            let db_operation = attributes
-                .get_value(consts::DB_OPERATION)
-                .and_then(|v| v.as_str());
-
-            let collection_name = attributes
-                .get_value(consts::DB_COLLECTION_NAME)
-                .and_then(|v| v.as_str());
-
-            let span_origin = attributes
-                .get_value(consts::ORIGIN)
-                .and_then(|v| v.as_str());
-
-            let (normalized_db_query, parsed_sql) = if let Some(raw_query) = raw_query {
-                scrub_db_query(
-                    raw_query,
-                    sub_op,
-                    db_system,
-                    db_operation,
-                    collection_name,
-                    span_origin,
-                )
-            } else {
-                (None, None)
-            };
-
-            let db_operation = if db_operation.is_none() {
-                if sub_op == "redis" || db_system == Some("redis") {
-                    // This only works as long as redis span descriptions contain the command + " *"
-                    if let Some(query) = normalized_db_query.as_ref() {
-                        let command = query.replace(" *", "");
-                        if command.is_empty() {
-                            None
-                        } else {
-                            Some(command)
-                        }
-                    } else {
-                        None
-                    }
-                } else if let Some(raw_query) = raw_query {
-                    // For other database operations, try to get the operation from data
-                    sql_action_from_query(raw_query).map(|a| a.to_uppercase())
-                } else {
-                    None
-                }
-            } else {
-                db_operation.map(|db_operation| db_operation.to_uppercase())
-            };
-
-            let db_collection_name: Option<String> = if collection_name.is_none() {
-                if span_origin == Some("auto.db.supabase") {
-                    normalized_db_query
-                        .as_ref()
-                        .and_then(|query| query.strip_prefix("from("))
-                        .and_then(|s| s.strip_suffix(")"))
-                        .map(String::from)
-                } else if let Some(raw_query) = raw_query {
-                    sql_tables_from_query(raw_query, &parsed_sql)
-                } else {
-                    None
-                }
-            } else if let Some(collection_name) = collection_name
-                && db_system == Some("mongodb")
-            {
-                if let Cow::Owned(s) = TABLE_NAME_REGEX.replace_all(collection_name, "{%s}") {
-                    Some(s)
-                } else {
-                    Some(collection_name.to_owned())
-                }
-            } else {
-                None
-            };
-
-            (normalized_db_query, db_operation, db_collection_name)
-        })
-        .unwrap_or_default();
-
-    if let Some(attributes) = attributes.value_mut() {
-        if let Some(normalized_db_query) = normalized_db_query {
-            attributes.insert(consts::NORMALIZED_DB_QUERY, normalized_db_query);
-        }
-        if let Some(db_operation_name) = db_operation {
-            attributes.insert(consts::DB_OPERATION, db_operation_name)
-        }
-        if let Some(db_collection_name) = db_collection_name {
-            attributes.insert(consts::DB_COLLECTION_NAME, db_collection_name);
-        }
-    }
 }
 
 /// Validates the start and end timestamps of a span.
@@ -803,285 +671,5 @@ mod tests {
             ..Default::default()
         });
         assert!(r.is_err());
-    }
-    #[test]
-    fn test_normalize_span_infers_op() {
-        let mut span = Annotated::<SpanV2>::from_json(
-            r#"
-        {
-            "start_timestamp": 1544719859.0,
-            "end_timestamp": 1544719860.0,
-            "trace_id": "5b8efff798038103d269b633813fc60c",
-            "span_id": "eee19b7ec3c1b174",
-            "name": "db.query",
-            "status": "ok",
-            "attributes": {
-                "db.system.name": {
-                    "type": "string",
-                    "value": "mysql"
-                },
-                "db.operation": {
-                    "type": "string",
-                    "value": "query"
-                }
-            }
-        }"#,
-        )
-        .unwrap();
-
-        let dsn = "https://e12d836b15bb49d7bbf99e64295d995b:@sentry.io/42"
-            .parse()
-            .unwrap();
-        let request_meta = crate::extractors::RequestMeta::new(dsn);
-        let envelope = crate::envelope::Envelope::from_request(None, request_meta);
-
-        normalize_span(
-            &mut span,
-            envelope.headers(),
-            &GeoIpLookup::default(),
-            make_context(relay_pii::DataScrubbingConfig::default(), None),
-        )
-        .unwrap();
-        assert_eq!(
-            span.value()
-                .unwrap()
-                .attributes
-                .value()
-                .unwrap()
-                .get_value(consts::OP)
-                .unwrap()
-                .as_str(),
-            Some("db")
-        );
-    }
-
-    #[test]
-    fn test_normalize_mysql_db_query_attributes() {
-        let mut attributes = Annotated::<Attributes>::from_json(
-            r#"
-        {
-          "sentry.op": {
-            "type": "string",
-            "value": "db.query"
-          },
-          "sentry.origin": {
-            "type": "string",
-            "value": "auto.otlp.spans"
-          },
-          "db.system.name": {
-            "type": "string",
-            "value": "mysql"
-          },
-          "db.query.text": {
-            "type": "string",
-            "value": "SELECT \"not an identifier\""
-          }
-        }
-        "#,
-        )
-        .unwrap();
-
-        normalize_db_attributes(&mut attributes);
-
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
-        {
-          "db.operation": {
-            "type": "string",
-            "value": "SELECT"
-          },
-          "db.query.text": {
-            "type": "string",
-            "value": "SELECT \"not an identifier\""
-          },
-          "db.system.name": {
-            "type": "string",
-            "value": "mysql"
-          },
-          "sentry.normalized_db_query": {
-            "type": "string",
-            "value": "SELECT %s"
-          },
-          "sentry.op": {
-            "type": "string",
-            "value": "db.query"
-          },
-          "sentry.origin": {
-            "type": "string",
-            "value": "auto.otlp.spans"
-          }
-        }
-        "#);
-    }
-
-    #[test]
-    fn test_normalize_mongodb_db_query_attributes() {
-        let mut attributes = Annotated::<Attributes>::from_json(
-            r#"
-        {
-          "sentry.op": {
-            "type": "string",
-            "value": "db"
-          },
-          "db.system.name": {
-            "type": "string",
-            "value": "mongodb"
-          },
-          "db.query.text": {
-            "type": "string",
-            "value": "{\"find\": \"documents\", \"foo\": \"bar\"}"
-          },
-          "db.operation": {
-            "type": "string",
-            "value": "find"
-          },
-          "db.collection.name": {
-            "type": "string",
-            "value": "documents"
-          }
-        }
-        "#,
-        )
-        .unwrap();
-
-        normalize_db_attributes(&mut attributes);
-
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
-        {
-          "db.collection.name": {
-            "type": "string",
-            "value": "documents"
-          },
-          "db.operation": {
-            "type": "string",
-            "value": "FIND"
-          },
-          "db.query.text": {
-            "type": "string",
-            "value": "{\"find\": \"documents\", \"foo\": \"bar\"}"
-          },
-          "db.system.name": {
-            "type": "string",
-            "value": "mongodb"
-          },
-          "sentry.normalized_db_query": {
-            "type": "string",
-            "value": "{\"find\":\"documents\",\"foo\":\"?\"}"
-          },
-          "sentry.op": {
-            "type": "string",
-            "value": "db"
-          }
-        }
-        "#);
-    }
-
-    #[test]
-    fn test_normalize_db_attributes_does_not_update_attributes_if_already_normalized() {
-        let mut attributes = Annotated::<Attributes>::from_json(
-            r#"
-        {
-          "db.collection.name": {
-            "type": "string",
-            "value": "documents"
-          },
-          "db.operation": {
-            "type": "string",
-            "value": "FIND"
-          },
-          "db.query.text": {
-            "type": "string",
-            "value": "{\"find\": \"documents\", \"foo\": \"bar\"}"
-          },
-          "db.system.name": {
-            "type": "string",
-            "value": "mongodb"
-          },
-          "sentry.normalized_db_query": {
-            "type": "string",
-            "value": "{\"find\":\"documents\",\"foo\":\"?\"}"
-          },
-          "sentry.op": {
-            "type": "string",
-            "value": "db"
-          }
-        }
-        "#,
-        )
-        .unwrap();
-
-        normalize_db_attributes(&mut attributes);
-
-        insta::assert_json_snapshot!(
-            SerializableAnnotated(&attributes), @r#"
-        {
-          "db.collection.name": {
-            "type": "string",
-            "value": "documents"
-          },
-          "db.operation": {
-            "type": "string",
-            "value": "FIND"
-          },
-          "db.query.text": {
-            "type": "string",
-            "value": "{\"find\": \"documents\", \"foo\": \"bar\"}"
-          },
-          "db.system.name": {
-            "type": "string",
-            "value": "mongodb"
-          },
-          "sentry.normalized_db_query": {
-            "type": "string",
-            "value": "{\"find\":\"documents\",\"foo\":\"?\"}"
-          },
-          "sentry.op": {
-            "type": "string",
-            "value": "db"
-          }
-        }
-        "#
-        );
-    }
-
-    #[test]
-    fn test_normalize_db_attributes_does_not_change_non_db_spans() {
-        let mut attributes = Annotated::<Attributes>::from_json(
-            r#"
-        {
-          "sentry.op": {
-            "type": "string",
-            "value": "http.client"
-          },
-          "sentry.origin": {
-            "type": "string",
-            "value": "auto.otlp.spans"
-          },
-          "http.request.method": {
-            "type": "string",
-            "value": "GET"
-          }
-        }
-      "#,
-        )
-        .unwrap();
-
-        normalize_db_attributes(&mut attributes);
-
-        insta::assert_json_snapshot!(SerializableAnnotated(&attributes), @r#"
-        {
-          "http.request.method": {
-            "type": "string",
-            "value": "GET"
-          },
-          "sentry.op": {
-            "type": "string",
-            "value": "http.client"
-          },
-          "sentry.origin": {
-            "type": "string",
-            "value": "auto.otlp.spans"
-          }
-        }
-        "#);
     }
 }
